+++ src/VBox/VMM/VMMR3/PDMAsyncCompletionFileNormal.cpp
@@ -1,4 +1,4 @@
-/* $Id: PDMAsyncCompletionFileNormal.cpp 106320 2024-10-15 12:08:41Z klaus.espenlaub@oracle.com $ */
+/* $Id: PDMAsyncCompletionFileNormal.cpp 108651 2025-03-06 11:15:45Z alexander.eichner@oracle.com $ */
 /** @file
  * PDM Async I/O - Async File I/O manager.
  */
@@ -998,7 +998,7 @@ static int pdmacFileAioMgrNormalProcessTaskList(PPDMACTASKFILE pTaskHead,
         RTMSINTERVAL msWhenNext;
         PPDMACTASKFILE pCurr = pTaskHead;
 
-        if (!pdmacEpIsTransferAllowed(&pEndpoint->Core, (uint32_t)pCurr->DataSeg.cbSeg, &msWhenNext))
+        if (!pdmacEpIsTransferAllowed(&pEndpoint->Core, pCurr->DataSeg.cbSeg, &msWhenNext))
         {
             pAioMgr->msBwLimitExpired = RT_MIN(pAioMgr->msBwLimitExpired, msWhenNext);
             break;
